+++ lib/std/js/jshttpclient.nim
@@ -1,0 +1,5 @@
+## HttpClient for JavaScript targets.
+## * Asyncronous `HttpClient` implemented on top of `jsfetch`.
+## * Syncronous `HttpClient` implemented on top of `jsxmlhttprequest`.
+import std/js/[jsasynchttpclient, jssynchttpclient]
+export jsasynchttpclient, jssynchttpclient