+++ tools/kochdocs.nim
@@ -15,7 +15,10 @@ const
   webUploadOutput = "web/upload"
 
 var nimExe*: string
-const allowList = ["jsbigints.nim", "jsheaders.nim", "jsformdata.nim", "jsfetch.nim", "jsutils.nim"]
+const allowList = [
+  "jsbigints.nim", "jsheaders.nim", "jsformdata.nim", "jsfetch.nim", "jshttpclient.nim",
+  "jssynchttpclient.nim", "jsasynchttpclient.nim", "jsxmlhttprequest.nim", "jsutils.nim",
+]
 
 template isJsOnly(file: string): bool =
   file.isRelativeTo("lib/js") or